Refactor initSession function
