+++ hie-bios.cabal
@@ -23,6 +23,7 @@ Library
   Exposed-Modules:      HIE.Bios
                         HIE.Bios.Config
                         HIE.Bios.Cradle
+                        HIE.Bios.Environment
                         HIE.Bios.Debug
                         HIE.Bios.Flags
                         HIE.Bios.Types
@@ -49,7 +50,7 @@ Library
                         file-embed           >= 0.0.11 && < 0.1,
                         ghc                  >= 8.4.1 && < 8.7,
                         transformers         >= 0.5.5 && < 0.6,
-                        temporary            >= 1.3 && < 1.4,
+                        temporary            >= 1.2 && < 1.4,
                         text                 >= 1.2.3 && < 1.3,
                         unix-compat          >= 0.5.2 && < 0.6,
                         unordered-containers >= 0.2.10 && < 0.3,
